Optimize the search for the objects and tiles in map instead of iterate over all the tiles for drawing

+++ app/src/main/java/com/bapplications/maplemobile/gameplay/physics/Physics.java
@@ -96,9 +96,7 @@ public class Physics {
     }
 
     public Point getYBelow(Point position) {
-        short ground = footholdtree.getYBelow(position);
-
-//        return new Point(position.x, ground - 1);
+        float ground = footholdtree.getYBelow(position);
         return new Point(position.x, ground + 1);
     }
